change integer type to deal with shortcoming of Frama-C

+++ Reports/D4.2/D4.2.2-VV-Implementation/Bitwalker/Modified3/Bitwalker_peek.c
@@ -31,9 +31,11 @@ uint64_t Bitwalker_Peek(unsigned int Start,
   {
     unsigned int pos = Start + i;
 
-    unsigned int bit_index  = inverse_modulo(pos, 8);
-    uint8_t shift = Bitstream[pos / 8] >> bit_index;
-    uint8_t bit_as_byte = shift & 1;
+    unsigned int bit_index  = inverse_modulo(pos, 8); 
+
+    // treat as unsigned int for Frama-C
+    unsigned int shift = Bitstream[pos / 8] >> bit_index;
+    unsigned int bit_as_byte = shift & 1;
     //@ assert bit_as_byte == 0 || bit_as_byte == 1;
 
     retval = 2 * retval + bit_as_byte;
